requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
Pillow==8.1.0
pycodestyle==2.6.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
